improve reflection and increase test coverage
