+++ jsf-spring-boot-autoconfigure/src/test/java/org/joinfaces/configuration/ReflectiveServletContextConfigurerTest.java
@@ -16,20 +16,10 @@
 
 package org.joinfaces.configuration;
 
-import java.lang.annotation.RetentionPolicy;
-import java.util.Arrays;
-import java.util.Collections;
+import java.lang.reflect.Field;
 import java.util.List;
 
-import javax.servlet.ServletContext;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.junit.Before;
 import org.junit.Test;
-
-import org.springframework.mock.web.MockServletContext;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -38,152 +28,38 @@ import static org.assertj.core.api.Assertions.assertThat;
  */
 public class ReflectiveServletContextConfigurerTest {
 
-	private ReflectiveServletContextConfigurer servletContextConfigurer;
-	private ServletContext servletContext;
+	@Test
+	public void testResolveCollectionTypeRaw() throws NoSuchFieldException {
+		Field listRawClass = ReflectiveServletContextConfigurerTest.class.getDeclaredField("listRawClass");
 
-	@Before
-	public void setUp() {
-		this.servletContext = new MockServletContext();
+		Class<?> type = ReflectiveServletContextConfigurer.resolveCollectionItemType(listRawClass);
 
-		this.servletContextConfigurer = new ReflectiveServletContextConfigurer<TestProperties>(this.servletContext, new TestProperties());
+		assertThat(type).isEqualTo(Class.class);
 	}
 
 	@Test
-	public void testString() {
-		this.servletContextConfigurer.configure();
+	public void testResolveCollectionTypeGeneric() throws NoSuchFieldException {
+		Field listGenericClass = ReflectiveServletContextConfigurerTest.class.getDeclaredField("listGenericClass");
 
-		assertThat(this.servletContext.getInitParameter("test.STRING")).isEqualTo("fooBar");
+		Class<?> type = ReflectiveServletContextConfigurer.resolveCollectionItemType(listGenericClass);
+
+		assertThat(type).isEqualTo(Class.class);
 	}
 
 	@Test
-	public void testSuperString() {
-		this.servletContextConfigurer.configure();
+	public void testRawList() throws NoSuchFieldException {
+		Field listRaw = ReflectiveServletContextConfigurerTest.class.getDeclaredField("listRaw");
 
-		assertThat(this.servletContext.getInitParameter("test.superString")).isEqualTo("barFoo");
+		Class<?> type = ReflectiveServletContextConfigurer.resolveCollectionItemType(listRaw);
+
+		assertThat(type).isEqualTo(Object.class);
 	}
 
-	@Test
-	public void testEmptyList() {
-		this.servletContextConfigurer.configure();
+	@SuppressWarnings("unused")
+	private List listRaw;
+	@SuppressWarnings("unused")
+	private List<Class> listRawClass;
+	@SuppressWarnings("unused")
+	private List<Class<?>> listGenericClass;
 
-		assertThat(this.servletContext.getInitParameter("test.EMPTY_LIST")).isEmpty();
-	}
-
-	@Test
-	public void testStringList() {
-		this.servletContextConfigurer.configure();
-
-		assertThat(this.servletContext.getInitParameter("test.LIST")).isEqualTo("foo...bar");
-	}
-
-	@Test
-	public void testSingletonList() {
-		this.servletContextConfigurer.configure();
-
-		assertThat(this.servletContext.getInitParameter("test.SINGLETON_LIST")).isEqualTo("foo");
-	}
-
-	@Test
-	public void testNull() {
-		this.servletContextConfigurer.configure();
-
-		assertThat(this.servletContext.getInitParameter("test.NULL")).isNull();
-	}
-
-	@Test
-	public void testNestedProperty() {
-		this.servletContextConfigurer.configure();
-
-		assertThat(this.servletContext.getInitParameter("test.inner.STRING")).isEqualTo("bar");
-	}
-
-	@Test
-	public void testClassList() {
-		this.servletContextConfigurer.configure();
-
-		assertThat(this.servletContext.getInitParameter("test.CLASS_LIST")).isEqualTo("java.lang.String;java.lang.Void");
-	}
-
-	@Test
-	public void testEnum() {
-		this.servletContextConfigurer.configure();
-
-		assertThat(this.servletContext.getInitParameter("test.ENUM")).isEqualTo("SOURCE");
-	}
-
-	@Test
-	public void testBoolean() {
-		this.servletContextConfigurer.configure();
-
-		assertThat(this.servletContext.getInitParameter("test.BOOLEAN")).isEqualTo("false");
-	}
-
-	@Test
-	public void testInt() {
-		this.servletContextConfigurer.configure();
-
-		assertThat(this.servletContext.getInitParameter("test.INT")).isEqualTo("42");
-	}
-
-	@Getter
-	@Setter
-	@NoArgsConstructor
-	public static class TestProperties extends SuperProperties {
-
-		@InitParameter("test.STRING")
-		private String testString = "fooBar";
-
-		@InitParameter(value = "test.EMPTY_LIST", listSeparator = "...")
-		private List<String> testEmptyList = Collections.emptyList();
-
-		@InitParameter(value = "test.LIST", listSeparator = "...")
-		private List<String> testStringList = Arrays.asList("foo", "bar");
-
-		@InitParameter(value = "test.SINGLETON_LIST", listSeparator = "...")
-		private List<String> singletonList = Collections.singletonList("foo");
-
-		@InitParameter("test.NULL")
-		private Object nullObject = null;
-
-		@NestedProperty
-		private InnerProperties innerProperties = new InnerProperties();
-
-		@NestedProperty
-		private InnerProperties nullInnerProperties = null;
-
-		@InitParameter(value = "test.CLASS_LIST", listSeparator = ";")
-		private List<Class<?>> classList = Arrays.asList(
-				String.class,
-				Void.class
-		);
-
-		private String notAnnotatedField;
-
-		@InitParameter("test.ENUM")
-		private RetentionPolicy testEnum = RetentionPolicy.SOURCE;
-
-		@InitParameter("test.BOOLEAN")
-		private boolean testBoolean = false;
-
-		@InitParameter("test.INT")
-		private int testInt = 42;
-
-		@Getter
-		@Setter
-		@NoArgsConstructor
-		public static class InnerProperties {
-
-			@InitParameter("test.inner.STRING")
-			private String innerString = "bar";
-		}
-	}
-
-	@Getter
-	@Setter
-	@NoArgsConstructor
-	public static class SuperProperties {
-
-		@InitParameter("test.superString")
-		private String superString = "barFoo";
-	}
 }
